feat: Tree support drag handler (#490)
* refactor: Move code into unique checker

* chore: support handler

* refactor: Move drag out

* chore: Patch more style info

* test: Fix test case

* docs: Update demo

* chore: Coverage
