+++ docs/examples/draggable.jsx
@@ -1,15 +1,32 @@
 /* eslint-disable no-console, react/no-access-state-in-setstate */
 import React from 'react';
-import { gData } from './utils/dataUtil';
+import { generateData } from './utils/dataUtil';
 import './draggable.less';
 import '../../assets/index.less';
-import Tree from 'rc-tree';
+import Tree from '../../src';
+
+const gData = generateData(2, 2, 2);
 
 class Demo extends React.Component {
   state = {
     gData,
     autoExpandParent: true,
-    expandedKeys: ['0-0-key', '0-0-0-key', '0-0-0-0-key'],
+    expandedKeys: [
+      '0-0-key',
+      '0-0-0-key',
+      '0-0-0-0-key',
+      '0-0-0-1-key',
+      '0-0-1-key',
+      '0-0-1-0-key',
+      '0-0-1-1-key',
+      '0-1-key',
+      '0-1-0-key',
+      '0-1-0-0-key',
+      '0-1-0-1-key',
+      '0-1-1-key',
+      '0-1-1-0-key',
+      '0-1-1-1-key',
+    ],
   };
 
   onDragStart = info => {
@@ -88,16 +105,21 @@ class Demo extends React.Component {
       <div className="draggable-demo">
         <h2>draggable</h2>
         <p>drag a node into another node</p>
-        <div className="draggable-container">
+        <div style={{ border: '1px solid red' }}>
           <Tree
             expandedKeys={this.state.expandedKeys}
             onExpand={this.onExpand}
             autoExpandParent={this.state.autoExpandParent}
             draggable
+            draggableIcon="↕️"
             onDragStart={this.onDragStart}
             onDragEnter={this.onDragEnter}
             onDrop={this.onDrop}
             treeData={this.state.gData}
+            // Virtual
+            height={200}
+            itemHeight={20}
+            virtual={false}
           />
         </div>
       </div>
